Avoid attempting to drop non-existant tables when downing a schema from a previous release and the new release has added a table

+++ lib/dbt/drivers/dialect/sql_server.rb
@@ -157,8 +157,9 @@ ORDER BY t.Ordinal, t.Name
         database_objects('SQL_INLINE_TABLE_VALUED_FUNCTION', schema_name).each { |name| execute("DROP FUNCTION #{name}") }
         database_objects('SQL_TABLE_VALUED_FUNCTION', schema_name).each { |name| execute("DROP FUNCTION #{name}") }
         database_objects('VIEW', schema_name).each { |name| execute("DROP VIEW #{name}") }
+        existing = database_objects('USER_TABLE', schema_name)
         tables.each do |table|
-          execute("DROP TABLE #{table}")
+          execute("DROP TABLE #{table}") if existing.include?(table)
         end
         database_objects('SEQUENCE_OBJECT', schema_name).each { |name| execute("DROP SEQUENCE #{name}") }
         execute("DROP SCHEMA #{schema_name}")
